requirements.txt
Flask==3.1.0
requests==2.32.0
python-dotenv==1.0.1
Pillow==11.0.0
gunicorn==23.0.0
pdf2image==1.16.3
psd-tools==1.9.27
pypdf2==3.0.1
celery==5.3.4
redis==5.0.1
opencv-python-headless==4.8.1.78
numpy==1.26.2
python-dateutil==2.8.2